Added a new utility class called RequestParamHelper
The idea of this class is to help with fetching possible parameters for API endpoints. It is geared toward POST requests and primarily consuming JSON body text.
This class should make it easier in the rest of the application to extract parameters from JSON objects sent to parameterized POST endpoints.