Fixed a bug that would cause a Route to crash with an IOException (#126)
* Fixed a bug that would cause a Route to crash with an IOException thrown by Jackson. If a document contained a field that a model did not define, the exception would be thrown.

This was fixed by setting a custom ObjectMapper in CollectionFactory. As a result, all JacksonMongoCollection objects now re-use an ObjectMapper (performance win!)
The ObjectMapper in ServiceLocator was also refactored to look cleaner, but functionally it doesn't change much. In the future, maybe a single ObjectMapper can be re-used. But for now, there are two instances of Jackson
being used to serialize/deserialize response models and database serialization/deserialization.

The MongoJack library was bumped from ver 4.0.0 to 4.0.1 since the newer version supplies a utility class to configure an ObjectMapper correctly.
The version bump results in a version bump for the Mongo driver from 4.0.0 to 4.0.3 .. Hopefully that doesn't break anything.